+++ apps/api/src/app/admin/admin.controller.ts
@@ -17,7 +17,8 @@ import {
   AdminData,
   AdminMarketData,
   AdminUsers,
-  EnhancedSymbolProfile
+  EnhancedSymbolProfile,
+  ScraperConfiguration
 } from '@ghostfolio/common/interfaces';
 import { permissions } from '@ghostfolio/common/permissions';
 import type {
@@ -222,13 +223,12 @@ export class AdminController {
   @Post('market-data/:dataSource/:symbol/test')
   @UseGuards(AuthGuard('jwt'), HasPermissionGuard)
   public async testMarketData(
-    @Body() data: { scraperConfiguration: string },
+    @Body() data: { scraperConfiguration: ScraperConfiguration },
     @Param('dataSource') dataSource: DataSource,
     @Param('symbol') symbol: string
   ): Promise<{ price: number }> {
     try {
-      const scraperConfiguration = JSON.parse(data.scraperConfiguration);
-      const price = await this.manualService.test(scraperConfiguration);
+      const price = await this.manualService.test(data.scraperConfiguration);
 
       if (price) {
         return { price };