+++ apps/api/src/app/user/user.service.ts
@@ -182,7 +182,7 @@ export class UserService {
       Access,
       accessToken,
       accounts,
-      Analytics,
+      analytics,
       authChallenge,
       createdAt,
       id,
@@ -198,7 +198,7 @@ export class UserService {
         accounts: {
           include: { platform: true }
         },
-        Analytics: true,
+        analytics: true,
         Settings: true,
         subscriptions: true
       },
@@ -217,9 +217,9 @@ export class UserService {
       Settings: Settings as UserWithSettings['Settings'],
       thirdPartyId,
       updatedAt,
-      activityCount: Analytics?.activityCount,
+      activityCount: analytics?.activityCount,
       dataProviderGhostfolioDailyRequests:
-        Analytics?.dataProviderGhostfolioDailyRequests
+        analytics?.dataProviderGhostfolioDailyRequests
     };
 
     if (user?.Settings) {
@@ -273,19 +273,27 @@ export class UserService {
       ).getSettings(user.Settings.settings),
       AssetClassClusterRiskEquity: new AssetClassClusterRiskEquity(
         undefined,
+        undefined,
+        undefined,
         undefined
       ).getSettings(user.Settings.settings),
       AssetClassClusterRiskFixedIncome: new AssetClassClusterRiskFixedIncome(
+        undefined,
+        undefined,
         undefined,
         undefined
       ).getSettings(user.Settings.settings),
       CurrencyClusterRiskBaseCurrencyCurrentInvestment:
         new CurrencyClusterRiskBaseCurrencyCurrentInvestment(
           undefined,
+          undefined,
+          undefined,
           undefined
         ).getSettings(user.Settings.settings),
       CurrencyClusterRiskCurrentInvestment:
         new CurrencyClusterRiskCurrentInvestment(
+          undefined,
+          undefined,
           undefined,
           undefined
         ).getSettings(user.Settings.settings),
@@ -380,7 +388,7 @@ export class UserService {
           frequency = 6;
         }
 
-        if (Analytics?.activityCount % frequency === 1) {
+        if (analytics?.activityCount % frequency === 1) {
           currentPermissions.push(permissions.enableSubscriptionInterstitial);
         }